+++ entidades/Obstaculos/Pilha.h
@@ -16,7 +16,7 @@ namespace Entidades {
 
         void executar();
         void obstaculizar(Personagens::Jogador* p);
-        void setPosicao(float novoX, float novoY);
+        void setPosicao(int novoX, int novoY);
         void salvar() ; /*fazer*/
   
     };
